libtest: extract stopwatch to a separate module
