+++ lib/logmsg/tests/test_logmsg_serialize.c
@@ -23,6 +23,7 @@
  */
 
 #include "testutils.h"
+#include "stopwatch.h"
 #include "apphook.h"
 #include "cfg.h"
 #include "plugin.h"
